Remove Duster config
