--- duster.json
@@ -1,3 +1,0 @@
-{
-    "exclude": ["tests/Unit/**"]
-}
