+++ tests/Unit/HasChildren/Foo.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Parental\Tests\Unit\HasChildren;
+
+class Foo
+{
+    //
+}
