+++ tests/Unit/HasChildren/HasChildrenChildModel.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Parental\Tests\Unit\HasChildren;
+
+class HasChildrenChildModel extends HasChildrenParentModel
+{
+    public $mutatorWasCalled = false;
+
+    public function setTestAttribute()
+    {
+        $this->mutatorWasCalled = true;
+    }
+}
