+++ tests/Unit/HasChildrenTest.php
@@ -2,9 +2,12 @@
 
 namespace Parental\Tests\Unit;
 
-use Illuminate\Database\Eloquent\Model;
-use Parental\HasChildren;
 use Parental\Tests\TestCase;
+use Parental\Tests\Unit\HasChildren\Bar;
+use Parental\Tests\Unit\HasChildren\Foo;
+use Parental\Tests\Unit\HasChildren\HasChildrenChildModel;
+use Parental\Tests\Unit\HasChildren\HasChildrenParentModel;
+use Parental\Tests\Unit\HasChildren\HasChildrenParentModelWithMethodTypes;
 
 class HasChildrenTest extends TestCase
 {
@@ -28,35 +31,5 @@ class HasChildrenTest extends TestCase
             'foo' => Foo::class,
             'bar' => Bar::class,
         ], $types);
-    }
-}
-
-class HasChildrenParentModel extends Model
-{
-    use HasChildren;
-
-    protected $fillable = ['type', 'test'];
-}
-
-class HasChildrenChildModel extends HasChildrenParentModel
-{
-    public $mutatorWasCalled = false;
-
-    public function setTestAttribute()
-    {
-        $this->mutatorWasCalled = true;
-    }
-}
-
-class HasChildrenParentModelWithMethodTypes extends Model
-{
-    use HasChildren;
-
-    public function getChildTypes()
-    {
-        return [
-            'foo' => Foo::class,
-            'bar' => Bar::class,
-        ];
     }
 }
