+++ tests/Unit/HasParent/ChildModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Parental\Tests\Unit\HasParent;
+
+use Parental\HasParent;
+
+class ChildModel extends ParentModel
+{
+    use HasParent;
+}
