+++ tests/Unit/HasParent/ChildModelWithoutTrait.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Parental\Tests\Unit\HasParent;
+
+class ChildModelWithoutTrait extends ParentModel
+{
+    //
+}
