+++ tests/Unit/HasParent/RelatedModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Parental\Tests\Unit\HasParent;
+
+use Illuminate\Database\Eloquent\Model;
+
+class RelatedModel extends Model
+{
+    //
+}
